Only depends on mescc for the Sinful creation and its dependence on mescc-tools is optional (simply replace the MESCC_TOOLS_PREFIX with ./)

+++ bootstrap.sh
@@ -1,3 +1,4 @@
+#! /usr/bin/env bash
 # Mes --- Maxwell Equations of Software
 # Copyright © 2017 Jan Nieuwenhuizen <janneke@gnu.org>
 # Copyright © 2017 Jeremiah Orians
@@ -17,17 +18,25 @@
 # You should have received a copy of the GNU General Public License
 # along with Mes.  If not, see <http://www.gnu.org/licenses/>.
 
-set -ex
+set -eux
+LIB_PREFIX=${LIB_PREFIX-./libs}
+MES_PREFIX=${MES_PREFIX-../mes}
+MESCC_TOOLS_PREFIX=${MESCC_TOOLS_PREFIX-../mescc-tools}
+
 # Sin that needs to be removed
 # Current hack to give us a holder for blood-elf.M1
-../mes/guile/mescc.scm -c -E -I ../mes/include -o blood-elf.E ../mescc-tools/blood-elf.c
-../mes/guile/mescc.scm -c -o blood-elf.M1 blood-elf.E
-# Current hack to give us a holder for M1.M1
-../mes/guile/mescc.scm -c -E -I ../mes/include -o M1.E ../mescc-tools/M1-macro.c
-../mes/guile/mescc.scm -c -o M1.M1 M1.E
-# Current hack to give us a holder for hex2.M1
-../mes/guile/mescc.scm -c -E -I ../mes/include -o hex2.E ../mescc-tools/hex2_linker.c
-../mes/guile/mescc.scm -c -o hex2.M1 hex2.E
+# To execute this block ./bootstrap.sh sin
+if [ "sin" = "${1-NOPE}" ]
+then
+	$MES_PREFIX/guile/mescc.scm -c -E -I $MES_PREFIX/include -o blood-elf.E $MESCC_TOOLS_PREFIX/blood-elf.c
+	$MES_PREFIX/guile/mescc.scm -c -o blood-elf.M1 blood-elf.E
+	# Current hack to give us a holder for M1.M1
+	$MES_PREFIX/guile/mescc.scm -c -E -I $MES_PREFIX/include -o M1.E $MESCC_TOOLS_PREFIX/M1-macro.c
+	$MES_PREFIX/guile/mescc.scm -c -o M1.M1 M1.E
+	# Current hack to give us a holder for hex2.M1
+	$MES_PREFIX/guile/mescc.scm -c -E -I $MES_PREFIX/include -o hex2.E $MESCC_TOOLS_PREFIX/hex2_linker.c
+	$MES_PREFIX/guile/mescc.scm -c -o hex2.M1 hex2.E
+fi
 
 #########################################
 # Phase-0 Build from external binaries  #
@@ -36,94 +45,102 @@ set -ex
 
 # blood-elf-0
 # Create proper debug segment
-../mescc-tools/bin/blood-elf -f blood-elf.M1 -o blood-elf-footer.M1
+$MESCC_TOOLS_PREFIX/bin/blood-elf \
+	-f blood-elf.M1\
+	-o blood-elf-footer.M1
 
 # Build
 # M1-macro phase
-../mescc-tools/bin/M1 \
+$MESCC_TOOLS_PREFIX/bin/M1 \
 	--LittleEndian\
 	--Architecture=1\
-	-f ../mes/stage0/x86.M1\
-	-f ../mes/lib/crt1.M1\
-	-f ../mes/lib/libc+tcc-mes.M1\
+	-f $LIB_PREFIX/x86.M1\
+	-f $LIB_PREFIX/crt1.M1\
+	-f $LIB_PREFIX/libc+tcc-mes.M1\
 	-f blood-elf.M1\
 	-f blood-elf-footer.M1\
 	-o blood-elf.hex2
+
 # Hex2-linker phase
-../mescc-tools/bin/hex2 \
+$MESCC_TOOLS_PREFIX/bin/hex2 \
 	--LittleEndian\
 	--Architecture=1\
 	--BaseAddress=0x1000000\
-	-f ../mes/stage0/elf32-header.hex2\
+	-f $LIB_PREFIX/elf32-header.hex2\
 	-f blood-elf.hex2\
 	-o blood-elf-0\
 	--exec_enable
 
 # Clean up temp files used in build
-rm blood-elf.E
-rm blood-elf.hex2
-rm blood-elf-footer.M1
+[ -f blood-elf.E ] && rm blood-elf.E
+[ -f blood-elf.hex2 ] && rm blood-elf.hex2
+[ -f blood-elf-footer.M1 ] && rm blood-elf-footer.M1
 
 # M1-0
 # Create proper debug segment
-../mescc-tools/bin/blood-elf -f M1.M1 -o M1-footer.M1
+$MESCC_TOOLS_PREFIX/bin/blood-elf \
+	-f M1.M1 \
+	-o M1-footer.M1
 
 # Build
 # M1-macro phase
-../mescc-tools/bin/M1 \
+$MESCC_TOOLS_PREFIX/bin/M1 \
 	--LittleEndian\
 	--Architecture=1\
-	-f ../mes/stage0/x86.M1\
-	-f ../mes/lib/crt1.M1\
-	-f ../mes/lib/libc+tcc-mes.M1\
+	-f $LIB_PREFIX/x86.M1\
+	-f $LIB_PREFIX/crt1.M1\
+	-f $LIB_PREFIX/libc+tcc-mes.M1\
 	-f M1.M1\
 	-f M1-footer.M1\
 	-o M1.hex2
+
 # Hex2-linker phase
-../mescc-tools/bin/hex2 \
+$MESCC_TOOLS_PREFIX/bin/hex2 \
 	--LittleEndian\
 	--Architecture=1\
 	--BaseAddress=0x1000000\
-	-f ../mes/stage0/elf32-header.hex2\
+	-f $LIB_PREFIX/elf32-header.hex2\
 	-f M1.hex2\
-	-f ../mes/stage0/elf32-footer-single-main.hex2\
 	-o M1-0\
 	--exec_enable
 
 # Clean up temp files used in build
-rm M1.E
-rm M1.hex2
-rm M1-footer.M1
+[ -f M1.E ] && rm M1.E
+[ -f M1.hex2 ] && rm M1.hex2
+[ -f M1-footer.M1 ] && rm M1-footer.M1
 
 # hex2-0
 # Create proper debug segment
-../mescc-tools/bin/blood-elf -f hex2.M1 -o hex2-footer.M1
+$MESCC_TOOLS_PREFIX/bin/blood-elf \
+	-f hex2.M1\
+	-o hex2-footer.M1
 
 # Build
 # M1-macro phase
-../mescc-tools/bin/M1 \
+$MESCC_TOOLS_PREFIX/bin/M1 \
 	--LittleEndian\
 	--Architecture=1\
-	-f ../mes/stage0/x86.M1\
-	-f ../mes/lib/crt1.M1\
-	-f ../mes/lib/libc+tcc-mes.M1\
+	-f $LIB_PREFIX/x86.M1\
+	-f $LIB_PREFIX/crt1.M1\
+	-f $LIB_PREFIX/libc+tcc-mes.M1\
 	-f hex2.M1\
 	-f hex2-footer.M1\
 	-o hex2.hex2
+
 # Hex2-linker phase
-../mescc-tools/bin/hex2 \
+$MESCC_TOOLS_PREFIX/bin/hex2 \
 	--LittleEndian\
 	--Architecture=1\
 	--BaseAddress=0x1000000\
-	-f ../mes/stage0/elf32-header.hex2\
+	-f $LIB_PREFIX/elf32-header.hex2\
 	-f hex2.hex2\
 	-o hex2-0\
 	--exec_enable
 
 # Clean up temp files used in build
-rm hex2.E
-rm hex2.hex2
-rm hex2-footer.M1
+[ -f hex2.E ] && rm hex2.E
+[ -f hex2.hex2 ] && rm hex2.hex2
+[ -f hex2-footer.M1 ] && rm hex2-footer.M1
 
 
 #############################################
@@ -139,25 +156,26 @@ rm hex2-footer.M1
 ./M1-0 \
 	--LittleEndian\
 	--Architecture=1\
-	-f ../mes/stage0/x86.M1\
-	-f ../mes/lib/crt1.M1\
-	-f ../mes/lib/libc+tcc-mes.M1\
+	-f $LIB_PREFIX/x86.M1\
+	-f $LIB_PREFIX/crt1.M1\
+	-f $LIB_PREFIX/libc+tcc-mes.M1\
 	-f blood-elf.M1\
 	-f blood-elf-footer.M1\
 	-o blood-elf.hex2
+
 # Hex2-linker phase
 ./hex2-0 \
 	--LittleEndian\
 	--Architecture=1\
 	--BaseAddress=0x1000000\
-	-f ../mes/stage0/elf32-header.hex2\
+	-f $LIB_PREFIX/elf32-header.hex2\
 	-f blood-elf.hex2\
 	-o blood-elf\
 	--exec_enable
 
 # Clean up temp files used in build
-rm blood-elf.hex2
-rm blood-elf-footer.M1
+[ -f blood-elf.hex2 ] && rm blood-elf.hex2
+[ -f blood-elf-footer.M1 ] && rm blood-elf-footer.M1
 
 # M1
 # Create proper debug segment
@@ -168,9 +186,9 @@ rm blood-elf-footer.M1
 ./M1-0 \
 	--LittleEndian\
 	--Architecture=1\
-	-f ../mes/stage0/x86.M1\
-	-f ../mes/lib/crt1.M1\
-	-f ../mes/lib/libc+tcc-mes.M1\
+	-f $LIB_PREFIX/x86.M1\
+	-f $LIB_PREFIX/crt1.M1\
+	-f $LIB_PREFIX/libc+tcc-mes.M1\
 	-f M1.M1\
 	-f M1-footer.M1\
 	-o M1.hex2
@@ -179,15 +197,14 @@ rm blood-elf-footer.M1
 	--LittleEndian\
 	--Architecture=1\
 	--BaseAddress=0x1000000\
-	-f ../mes/stage0/elf32-header.hex2\
+	-f $LIB_PREFIX/elf32-header.hex2\
 	-f M1.hex2\
-	-f ../mes/stage0/elf32-footer-single-main.hex2\
 	-o M1\
 	--exec_enable
 
 # Clean up temp files used in build
-rm M1.hex2
-rm M1-footer.M1
+[ -f M1.hex2 ] && rm M1.hex2
+[ -f M1-footer.M1 ] && rm M1-footer.M1
 
 # hex2
 # Create proper debug segment
@@ -198,9 +215,9 @@ rm M1-footer.M1
 ./M1-0 \
 	--LittleEndian\
 	--Architecture=1\
-	-f ../mes/stage0/x86.M1\
-	-f ../mes/lib/crt1.M1\
-	-f ../mes/lib/libc+tcc-mes.M1\
+	-f $LIB_PREFIX/x86.M1\
+	-f $LIB_PREFIX/crt1.M1\
+	-f $LIB_PREFIX/libc+tcc-mes.M1\
 	-f hex2.M1\
 	-f hex2-footer.M1\
 	-o hex2.hex2
@@ -209,14 +226,14 @@ rm M1-footer.M1
 	--LittleEndian\
 	--Architecture=1\
 	--BaseAddress=0x1000000\
-	-f ../mes/stage0/elf32-header.hex2\
+	-f $LIB_PREFIX/elf32-header.hex2\
 	-f hex2.hex2\
 	-o hex2\
 	--exec_enable
 
 # Clean up temp files used in build
-rm hex2.hex2
-rm hex2-footer.M1
+[ -f hex2.hex2 ] && rm hex2.hex2
+[ -f hex2-footer.M1 ] && rm hex2-footer.M1
 
 #########################################
 # Phase-2 Check for unmatched output    #
@@ -232,6 +249,7 @@ diff -a hex2 hex2-0 || echo "hex2 MISMATCH"
 #########################################
 # Phase-3 Clean up files being tested   #
 #########################################
-rm M1-0
-rm blood-elf-0
-rm hex2-0
+[ -f M1-0 ] && rm M1-0
+[ -f blood-elf-0 ] && rm blood-elf-0
+[ -f hex2-0 ] && rm hex2-0
+echo "SUCCESS"
